requirements.txt
termcolor==2.2.0
python-dotenv==1.0.0
pymongo==4.3.3
requests==2.28.2